needed int before sizeof

+++ hello_world/6-size.c
@@ -1,5 +1,4 @@
 #include <stdio.h>
-#include <string.h>
 
 /**
  * main - Entry point
@@ -10,10 +9,10 @@
 int main(void)
 
 {
-	printf("Size of char: %lu byte(s)\n", sizeof(char));
-	printf("Size of an int: %lu byte(s)\n", sizeof(int));
-	printf("Size of a long int: %lu byte(s)\n", sizeof(long int));
-	printf("Size of a long long int: %lu byte(s)\n", sizeof(long long int));
-	printf("Size of a float: %lu byte(s)\n", sizeof(float));
+	printf("Size of char: %lu byte(s)\n", (int)sizeof(char));
+	printf("Size of an int: %lu byte(s)\n", (int)sizeof(int));
+	printf("Size of a long int: %lu byte(s)\n", (int)sizeof(long int));
+	printf("Size of a long long int: %lu byte(s)\n", (int)sizeof(long long int));
+	printf("Size of a float: %lu byte(s)\n", (int)sizeof(float));
 	return (0);
 }
